feat: OOP를 잘 지키기 위해서 DAO 계층 분리 및 리팩터링
#5
