+++ src/main/kotlin/swyg/hollang/repository/TestJpaRepository.kt
@@ -1,0 +1,13 @@
+package swyg.hollang.repository
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.data.jpa.repository.Query
+import swyg.hollang.entity.Test
+
+interface TestJpaRepository: JpaRepository<Test, Long> {
+
+    @Query("select t from Test t " +
+            "join fetch t.questions tq " +
+            "where t.version = :version")
+    fun findWithQuestionsAndAnswersByVersion(version: Long): Test?
+}
