+++ src/main/kotlin/swyg/hollang/repository/TestRepository.kt
@@ -1,13 +1,9 @@
 package swyg.hollang.repository
 
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.jpa.repository.Query
 import swyg.hollang.entity.Test
 
-interface TestRepository: JpaRepository<Test, Long> {
+interface TestRepository {
 
-    @Query("select t from Test t " +
-            "join fetch t.questions tq " +
-            "where t.version = :version")
-    fun findWithQuestionsAndAnswersByVersion(version: Long): Test?
+    fun findWithQuestionsAndAnswersByVersion(version: Long): Test
+
 }
