--- src/main/kotlin/swyg/hollang/service/ShuffleService.kt
@@ -1,10 +1,0 @@
-package swyg.hollang.service
-
-import swyg.hollang.entity.Question
-
-class ShuffleService {
-
-    fun shuffleElements(questions: MutableList<Question>) {
-        questions.shuffle()
-    }
-}
